backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pg8000==1.30.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic v2 (with pre-built wheels for Python 3.12)
pydantic==2.5.3
pydantic-settings==2.1.0

# Groq AI
groq==0.4.2

# HTTP client
httpx==0.26.0

# Rate limiting
slowapi==0.1.9

# Retry logic
tenacity==8.2.3

# Logging
structlog==24.1.0

# Environment
python-dotenv==1.0.0
